Splitted components into different files
