+++ assets/css/style.css
@@ -67,7 +67,8 @@ li a {
     width:36px;
     height:36px;
     cursor: pointer;
-    margin-top: 5px;
+    /*margin-top: 5px;*/
+    margin: 5px;
 }
 .contact_icon {
     transition: all 0.3s linear;
